% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anidar-horizontes.R
\name{anidar_horizontes}
\alias{anidar_horizontes}
\title{Anida datos de horizontes}
\usage{
anidar_horizontes(perfiles)
}
\arguments{
\item{perfiles}{un data.frame con perfiles (salida de \code{\link[=get_perfiles]{get_perfiles()}})}
}
\value{
Un data.frame anidado. El rsultado tiene una fila por cada perfil y una
columna por cada dato del perfil más una columna llamada "horizontes" que
tiene un data.frame con los horizontes de ese perfil.
}
\description{
Anida datos de horizontes
}
\examples{

perfiles <- get_perfiles(c(3238, 4634))
anidar_horizontes(perfiles)

}
