% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-inta.R
\name{check_sisinta}
\alias{check_sisinta}
\title{Evalúa conexión con el servidor de SISINTA}
\usage{
check_sisinta()
}
\value{
Un valor lógico indicando si se puede establecer conexión con el servidor
}
\description{
Evalúa conexión con el servidor de SISINTA
}
