% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar-shapefile.R
\name{exportar_shapefile}
\alias{exportar_shapefile}
\title{Exportar perfiles en shapefile}
\usage{
exportar_shapefile(perfiles, archivo)
}
\arguments{
\item{perfiles}{un data.frame con perfiles (salida de \code{\link[=get_perfiles]{get_perfiles()}})}

\item{archivo}{la ruta al archivo de salida.}
}
\description{
Exportar perfiles en shapefile
}
\examples{
perfiles <- get_perfiles(c(3238, 4634))
archivo <- tempfile(fileext = ".shp")
# exportar_shapefile(perfiles, archivo)

}
