% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-perfil.R
\name{get_perfiles}
\alias{get_perfiles}
\title{Obtiene perfiles de suelo}
\usage{
get_perfiles(perfil_ids, dir = tempdir(), refresh = FALSE)
}
\arguments{
\item{perfil_ids}{vector numérico de ids de perfiles de suelo.}

\item{dir}{directorio donde se guardan los datos de perfiles o
donde se leen perfiles ya descargados. Por defecto, los perfiles
descargados se guardan en una carpeta temporal.}

\item{refresh}{boleano, fuerza la descarga de los datos para
actualizar los perfiles ya descargados.}
}
\value{
Un data.frame.
}
\description{
\code{get_perfies} descarga perfiles desde SISINTA o lee perfiles
descargados previamente.
}
\examples{

get_perfiles(c(3238, 4634))


}
