% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar-na.R
\name{imputar_profundidad_inferior}
\alias{imputar_profundidad_inferior}
\title{Imputa valores faltantes en profundidad inferiror}
\usage{
imputar_profundidad_inferior(perfiles, profundidad = 5)
}
\arguments{
\item{perfiles}{un data.frame con perfiles (salida de \code{\link[=get_perfiles]{get_perfiles()}})}

\item{profundidad}{la profundidad asumida de la útlima capa}
}
\value{
Un data.frame igual al de entrada.
}
\description{
Imputa valores faltantes en profundidad inferiror
}
