% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolar-spline.R, R/promedio-ponderado.R
\name{interpolar_spline}
\alias{interpolar_spline}
\alias{interpolar_promedio_ponderado}
\title{Métodos de interpolación}
\usage{
interpolar_spline(lambda = 0.1)

interpolar_promedio_ponderado()
}
\arguments{
\item{lambda}{parámetro lambda de suavizado.}
}
\description{
Métodos de interpolación
}
