% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-in.R
\name{log_in}
\alias{log_in}
\title{Inicial sesión en la página de SISINTA}
\usage{
log_in(usuario, pass)
}
\arguments{
\item{usuario}{string, usualmente el mail.}

\item{pass}{string, contraseña asociada.}
}
\description{
Esta función interna es utilizada por \code{\link[=get_perfiles]{get_perfiles()}} para acceder a la web de
SISINTA con usuario y contraseña y poder descargar los perfiles no púplicos.
}
