% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolar-concatenar.R, R/interpolar-spline.R,
%   R/promedio-ponderado.R
\name{interpolar_concatenar}
\alias{interpolar_concatenar}
\alias{interpolar_spline}
\alias{interpolar_promedio_ponderado}
\title{Métodos de interpolación}
\usage{
interpolar_concatenar(sep = "|")

interpolar_spline(lambda = 0.1)

interpolar_promedio_ponderado()
}
\arguments{
\item{sep}{separador utilizado para concatenar los valores únicos de cada capa.}

\item{lambda}{parámetro lambda de suavizado.}
}
\description{
SISINTAR implementa dos métodos de interpolación para obtener horizontes estandarizados
de las distintas variables disponibles. Estos métodos devuelven una función que recibe
los límites superior e inferior de cada capa, las observaciones a interpolar y los
horizontes o profundidades a los que se quiere interpolar cada variable.
}
\details{
\code{interpolar_promedio_ponderado} utiliza un promedio ponderado para calcular los valores
interpolados en las capas definidas.

\code{interpolar_spline} utiliza la funcione spline que preserva la masa de acuerdo a
Bishop T.F.A. el. al. (1999).
}
\references{
Bishop T.F.A. el. al. (1999) Modelling soil attribute depth functions with equal-area quadratic
smoothing splines. \url{https://doi.org/10.1016/S0016-7061(99)00003-8}
}
