% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfiles-ejemplo.R
\docType{data}
\name{perfiles}
\alias{perfiles}
\title{Perfiles de ejemplo}
\format{
An object of class \code{data.frame} with 18 rows and 105 columns.
}
\usage{
perfiles
}
\description{
Perfiles de ejemplo
}
\keyword{datasets}
