% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separar-perfiles.R
\name{separar_perfiles}
\alias{separar_perfiles}
\title{Separar datos de sitio y horizonte en dos tablas}
\usage{
separar_perfiles(perfiles)
}
\arguments{
\item{perfiles}{un data.frame con perfiles (salida de \code{\link[=get_perfiles]{get_perfiles()}})}
}
\value{
Una lista con dos elementos. El elemento \code{sitios}, es un data.frmae con
la información de los sitios y el elemneto \code{horizontes} es un data.frame
con la información de horizontes.
}
\description{
Separar datos de sitio y horizonte en dos tablas
}
