% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis-perfi.R
\name{sis_perfiles}
\alias{sis_perfiles}
\title{Obtiene perfiles de suelo}
\usage{
sis_perfiles(perfil_ids, dir = tempdir())
}
\arguments{
\item{perfil_ids}{vector de id numéricos de perfiles de suelo.}

\item{dir}{directorio donde descargar los archivos.}
}
\value{
Un data.frame.
}
\description{
Obtiene perfiles de suelo
}
\examples{

sis_perfiles(c(3238, 4634))


}
