% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-soilprofilecolection.R
\name{as_SoilProfileCollection}
\alias{as_SoilProfileCollection}
\title{Objeto SoilProfileCollection}
\usage{
as_SoilProfileCollection(perfiles)
}
\arguments{
\item{perfiles}{data.frame con perfiles descargados o leidos con
la función \code{\link[=get_perfiles]{get_perfiles()}}.}
}
\value{
un objeto \link[aqp:SoilProfileCollection-class]{aqp::SoilProfileCollection}
}
\description{
Transforma uno másperfiles de SISINTA en formato data.frame en un objeto
SoilProfileCollection del paquete \code{aqp}.
}
\examples{

coleccion <- as_SoilProfileCollection(perfiles)

}
