% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar-excel.R
\name{exportar_excel}
\alias{exportar_excel}
\title{Exporta perfiles a excel}
\usage{
exportar_excel(perfiles, archivo)
}
\arguments{
\item{perfiles}{un data.frame con perfiles (salida de \code{\link[=get_perfiles]{get_perfiles()}})}

\item{archivo}{la ruta al archivo de salida.}
}
\value{
Devuelve la ruta al archivo de forma invisible.
}
\description{
Exporta un data.frame con datos de perfiles a un archivo excel donde los datos
de sitio están en una hoja llamada "sitios" y los horizontes, en una hoja
llamada "horizontes".
}
\examples{

archivo <- tempfile(fileext = ".xlsx")
exportar_excel(get_perfiles(c(3238, 4634)), archivo)

}
