% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-perfil.R
\name{get_perfiles}
\alias{get_perfiles}
\title{Obtiene perfiles de suelo}
\usage{
get_perfiles(
  perfil_ids,
  dir = tempdir(),
  refresh = FALSE,
  parar_en_error = FALSE,
  credenciales = NULL
)
}
\arguments{
\item{perfil_ids}{vector numérico de ids de perfiles de suelo.
Alternativamente puede ser un data.frame con una columna llamada
\code{perfil_id} (la salida de \code{\link[=buscar_perfiles]{buscar_perfiles()}})}

\item{dir}{directorio donde se guardan los datos de perfiles o
donde se leen perfiles ya descargados. Por defecto, los perfiles
descargados se guardan en una carpeta temporal.}

\item{refresh}{boleano, fuerza descargar datos de SISINTA.}

\item{parar_en_error}{tirar un error si algún perfil no está disponible
en vez de seguir intentando con los siguientes.}

\item{credenciales}{una lista con elementos "usuario" y "pass".}
}
\value{
Un data.frame.
}
\description{
Obtiene datos de perfiles.
Por defecto, primero intenta leer datos existentes en \code{dir}, si no existen,
usa los datos incluidos en el paquete, y si no existen, trata de descargarlos
desde SISINTA. Esto garantiza reproducibilidad del proyecto o bajo mismas
versiones del paquete y minimiza la dependencia de una conexión a internet.
}
\examples{

\dontrun{
get_perfiles(c(3238, 4634, 4609))
get_perfiles(4609, credenciales = list(usuario = "usuario",
                                       pass = "pass"))
}

}
