% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-perfil.R
\name{get_perfiles}
\alias{get_perfiles}
\title{Obtiene perfiles de suelo}
\usage{
get_perfiles(
  perfil_ids,
  dir = tempdir(),
  refresh = FALSE,
  parar_en_error = FALSE
)
}
\arguments{
\item{perfil_ids}{vector numérico de ids de perfiles de suelo.
Alternativamente puede ser un data.frame con una columna llamada
\code{perfil_id} (la salida de \code{\link[=buscar_perfiles]{buscar_perfiles()}})}

\item{dir}{directorio donde se guardan los datos de perfiles o
donde se leen perfiles ya descargados. Por defecto, los perfiles
descargados se guardan en una carpeta temporal.}

\item{refresh}{boleano, fuerza la descarga de los datos para
actualizar los perfiles ya descargados.}

\item{parar_en_error}{tirar un error si algún perfil no está disponible
seguir intentando con los siguientes.}
}
\value{
Un data.frame.
}
\description{
\code{get_perfies} descarga perfiles desde SISINTA o lee perfiles
descargados previamente.
}
\examples{

get_perfiles(c(3238, 4634))


}
