% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar-profundidad-inferior.R
\name{imputar_profundidad_inferior}
\alias{imputar_profundidad_inferior}
\title{Imputa valores faltantes en profundidad inferior}
\usage{
imputar_profundidad_inferior(perfiles, profundidad = 5)
}
\arguments{
\item{perfiles}{un data.frame con perfiles (salida de \code{\link[=get_perfiles]{get_perfiles()}})}

\item{profundidad}{la profundidad asumida de la ultima capa}
}
\value{
Un data.frame con las mismas filas y columnas que el objeto de entrada.
}
\description{
La función revisa si el perfil tiene valor faltante en la variable
\code{profundidad_inferior} e imputa un valor dado por \code{profundidad_superior}
más una constante definida por el argumento \code{profundidad}.
}
\examples{
perfil <- subset(perfiles, perfil_id == 3238)
imputar_profundidad_inferior(perfil)
imputar_profundidad_inferior(perfil, profundidad = 10)

}
