% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolar-perfiles.R
\name{interpolar_perfiles}
\alias{interpolar_perfiles}
\title{Interpola perfiles}
\usage{
interpolar_perfiles(perfiles, variables, horizontes = 30)
}
\arguments{
\item{perfiles}{un data.frame con datos de perfiles}

\item{variables}{un vector de texto con los nombres de las variables
a interpolar}

\item{horizontes}{un vector numérico que determina los horizontes a
usar para la interpolación o un numérico único que determina la resolución
en centímetros.}
}
\value{
Un data.frame con los datos interpolados.
}
\description{
La función genera perfiles normalizados a horizontes estandarizados
calculando un promedio ponderado.
}
\examples{
interpolar_perfiles(get_perfiles(c(3238, 4634)), c("analitico_s", "analitico_t"))
# Horizontes cada 10 centímetros entre 0 y 100.
interpolar_perfiles(get_perfiles(c(3238, 4634)), c("analitico_s", "analitico_t"), seq(0, 100, 10))

}
