% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separar-perfiles.R
\name{separar_perfiles}
\alias{separar_perfiles}
\title{Separar datos de sitio y horizonte en dos tablas}
\usage{
separar_perfiles(perfiles)
}
\arguments{
\item{perfiles}{un data.frame con perfiles (salida de \code{\link[=get_perfiles]{get_perfiles()}})}
}
\value{
Una lista con dos elementos. El elemento \code{sitios}, es un data.frame con
la informaci贸n de los sitios y el elemneto \code{horizontes} es un data.frame
con la informaci贸n de horizontes.
}
\description{
Esta funci贸n permite separar el data.frame que genera \code{\link[=get_perfiles]{get_perfiles()}} que contiene
la informaci贸n de perfiles y sus horizontes en dos data.frames organizados en una lista.
}
\examples{
perfiles <- get_perfiles(c(3238))
separar_perfiles(perfiles)

}
